+++ src/Delizious.Toolkit/Match.cs
@@ -579,6 +579,64 @@
         internal static IMatch<T> Any<T>(params IMatch<T>[] matches)
             => CompositeMatch<T>.Any(matches);
 
+        /// <summary>
+        /// Creates a <see cref="Match{T}"/> instance that matches successfully when a value matches none of the specified <paramref name="matches"/>.
+        /// </summary>
+        /// <typeparam name="T">
+        /// The type of the value to match.
+        /// </typeparam>
+        /// <param name="matches">
+        /// The matches a value must match none to match successfully.
+        /// </param>
+        /// <returns>
+        /// A new <see cref="Match{T}"/> instance that matches successfully when a value matches none of the specified <paramref name="matches"/>.
+        /// </returns>
+        /// <exception cref="ArgumentNullException">
+        /// <paramref name="matches"/> is <c>null</c>.
+        /// </exception>
+        /// <exception cref="ArgumentException">
+        /// <paramref name="matches"/> contain at least one match that is <c>null</c>.
+        /// </exception>
+        public static Match<T> None<T>(params Match<T>[] matches)
+        {
+            if (ReferenceEquals(matches, null!))
+            {
+                throw new ArgumentNullException(nameof(matches));
+            }
+
+            if (matches.Any(match => ReferenceEquals(match, null)))
+            {
+                throw new ArgumentException("At least one match is a null reference.", nameof(matches));
+            }
+
+            return Match<T>.None(matches);
+        }
+
+        internal static IMatch<T> None<T>(params IMatch<T>[] matches)
+            => NotMatch<T>.Create(CompositeMatch<T>.Any(matches));
+
+        /// <summary>
+        /// Creates a <see cref="Match{T}"/> instance that matches successfully when a value matches none of the specified <paramref name="matches"/>.
+        /// </summary>
+        /// <typeparam name="T">
+        /// The type of the value to match.
+        /// </typeparam>
+        /// <param name="matches">
+        /// The matches a value must match none to match successfully.
+        /// </param>
+        /// <returns>
+        /// A new <see cref="Match{T}"/> instance that matches successfully when a value matches none of the specified <paramref name="matches"/>.
+        /// </returns>
+        /// <exception cref="ArgumentNullException">
+        /// <paramref name="matches"/> is <c>null</c>.
+        /// </exception>
+        /// <exception cref="ArgumentException">
+        /// <paramref name="matches"/> contain at least one match that is <c>null</c>.
+        /// </exception>
+        [Obsolete("This method only exists for compatibility reasons and has been replaced by 'None' method due to better naming. It will be removed in an upcoming release.")]
+        public static Match<T> Except<T>(params Match<T>[] matches)
+            => None(matches);
+
         private sealed class CompositeMatch<T> : IMatch<T>
         {
             private delegate bool CompositeMatchDelegate(IEnumerable<IMatch<T>> matches, T value);
@@ -649,6 +707,9 @@
 
         internal static Match<T> Any(IEnumerable<Match<T>> matches)
             => Create(Match.Any(matches.Select(item => item.match).ToArray()));
+
+        internal static Match<T> None(IEnumerable<Match<T>> matches)
+            => Create(Match.None(matches.Select(item => item.match).ToArray()));
 
         /// <summary>
         /// Determines whether the specified <paramref name="value"/> successfully matches according to this match.